requirements.txt
absl-py==0.6.1
alabaster==0.7.10
alembic==1.0.0
aniso8601==4.1.0
AnyQt==0.0.8
appnope==0.1.0
appscript==1.0.1
apyori==1.1.1
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
Babel==2.5.3
bokeh==0.12.16
configparser==3.5.0
contextlib2==0.5.5
dataclasses==0.6
Django==2.0.2
django-filter==2.0.0
django-rest-swagger==2.2.0
django-suit==0.2.26
django-tables2==1.21.2
django-tastypie==0.14.1
djangorestframework==3.8.2
Flask==1.0.2
Flask-API==1.0
Flask-Compress==1.4.0
Flask-Cors==3.0.4
Flask-Migrate==2.2.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
html5lib==1.0.1
httplib2==0.12.0
Markdown==2.6.11
MarkupSafe==1.0
mathparse==0.1.2
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.3
msgpack==0.6.0
msgpack-python==0.5.6
nltk==3.3
numpy==1.17.2
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
pluggy==0.6.0
ply==3.11
preshed==2.0.1
prompt-toolkit==2.0.9
protobuf==3.6.1
psutil==5.4.5
psycopg2==2.7.5
ptyprocess==0.5.2
py==1.5.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
requests==2.18.4
requests-oauthlib==1.0.0
retrying==1.3.3
rope==0.10.7
rsa==4.0
ruamel-yaml==0.15.35
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.21.3
scipy==1.1.0
seaborn==0.9.0
sklearn==0.0
srsly==0.1.0
stanfordnlp==0.1.1
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
tensorboard==1.12.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
textblob==0.15.2
textmining==1.0
thinc==7.0.8
toolz==0.9.0
torch==1.0.0
torchvision==0.4.0
tornado==5.0.2
tqdm==4.36.1
traitlets==4.3.2
typing==3.6.4
umap==0.1.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
utils==0.9.0
virtualenv==16.0.0
websocket-client==0.54.0
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wordcloud==1.5.0
wrapt==1.10.11
xlrd==1.1.0
